requirements.txt
# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0